snn_pattern_learning/requirements.txt
# SNN Pattern Learning - Required Dependencies

# Core PyTorch
torch>=1.9.0
torchvision>=0.10.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0

# Configuration Management
PyYAML>=5.4.0

# Optional: For advanced functionality
# aihwkit>=0.5.0  # Uncomment if using analog AI hardware components